backend/requirements.txt
django==5.1.2
openai==1.51.2
python-dotenv==1.0.1
django-cors-headers==4.5.0
djangorestframework==3.15.2